requirements.txt
# E-Com Video Insider - Requirements
# Sprint 1: Data Pipeline

# Apify Client for video data extraction
apify-client==1.7.1

# Google Generative AI SDK for video analysis
google-generativeai==0.8.3

# Streamlit for UI
streamlit==1.39.0

# Utilities
python-dotenv==1.0.1
requests==2.31.0

# Video downloader
yt-dlp==2026.2.4
